server/requirements.txt
# Core dependencies
flask==3.0.0
flask-cors==4.0.0

# Google Calendar API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# OpenAI API
openai==1.98.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3

# Development dependencies
pytest==7.4.0
black==23.7.0
flake8==6.1.0
mypy==1.5.1